Updated SQLite with locations and restructured all DAO's and the helper

+++ ZoomCube/src/com/zoom/map/TagListener.java
@@ -1,7 +1,7 @@
 package com.zoom.map;
 
 public interface TagListener {
-	public void AddTag(String _tag, int _itemId);
-	public void DeleteTag(String _tag, int _itemId);
+	public void AddTag(String _tag, long _itemId);
+	public void DeleteTag(String _tag, long _itemId);
 	public void ReturnResults();
 }
